Update README.md
This is a communal repo by the MSc. Bioinformatics students where we gather our
code in a central place. If you wish to contribute, create a branch with the
name corresponding to your gitup username.
At some point we wish to consolidate the gathered code into ideal solutions to
the exercises